🚚 move[#5] : Channel Dto 디렉토리 이동
- infrastructure -> application 으로 이동
