+++ src/main/java/com/fouribnb/notification/application/dto/requestDto/ChannelRequest.java
@@ -1,4 +1,4 @@
-package com.fouribnb.notification.infrastructure.dto.requestDto;
+package com.fouribnb.notification.application.dto.requestDto;
 
 import java.util.UUID;
 import lombok.Builder;
