+++ src/main/java/com/fouribnb/notification/application/service/NotificationServiceImpl.java
@@ -6,8 +6,8 @@ import com.fouribnb.notification.application.mapper.ChannelMapper;
 import com.fouribnb.notification.application.mapper.NotificationMapper;
 import com.fouribnb.notification.domain.entity.Notification;
 import com.fouribnb.notification.domain.repository.NotificationRepository;
-import com.fouribnb.notification.infrastructure.dto.requestDto.ChannelRequest;
-import com.fouribnb.notification.infrastructure.dto.responseDto.ChannelResponse;
+import com.fouribnb.notification.application.dto.requestDto.ChannelRequest;
+import com.fouribnb.notification.application.dto.responseDto.ChannelResponse;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
