+++ src/main/java/com/fouribnb/notification/infrastructure/channel/SlackServiceImpl.java
@@ -4,8 +4,8 @@ import com.fouribnb.notification.application.mapper.ChannelMapper;
 import com.fouribnb.notification.application.service.ChannelService;
 import com.fouribnb.notification.domain.entity.Notification;
 import com.fouribnb.notification.domain.repository.NotificationRepository;
-import com.fouribnb.notification.infrastructure.dto.requestDto.ChannelRequest;
-import com.fouribnb.notification.infrastructure.dto.responseDto.ChannelResponse;
+import com.fouribnb.notification.application.dto.requestDto.ChannelRequest;
+import com.fouribnb.notification.application.dto.responseDto.ChannelResponse;
 import com.slack.api.Slack;
 import com.slack.api.methods.MethodsClient;
 import com.slack.api.methods.request.chat.ChatPostMessageRequest;
